Extract common command method
Previously removeNonValues was duplicated in lib/commands/mailbox.js and
lib/commands/folder.js - this patch extracts that function and puts it in
lib/commands/helpers/common.js
