+++ lib/commands/mailbox.js
@@ -12,21 +12,7 @@
 var dal;
 var logger;
 var sprintf = require('sprintf').sprintf;
-
-/**
- * Removes keys from an associative array that are either functions or objects
- *
- * @param {object} associativeArray - associative array to cull junk from
- */
-function removeNonValues(associativeArray) {
-  for (var key in associativeArray) {
-    var fieldType = typeof(associativeArray[key]);
-    if (fieldType === 'function' ||
-        fieldType === 'object') {
-      delete associativeArray[key];
-    }
-  }
-}
+var common = require('./helpers/common.js')();
 
 function createCommands(dependencies) {
   dal = dependencies.dal;
@@ -64,7 +50,7 @@ function createCommands(dependencies) {
                                     mailboxNumber, mailboxContext));
           }
 
-          removeNonValues(mailbox);
+          common.removeNonValues(mailbox);
           logger.info('%j', mailbox);
         });
     },
